panopto_summarizer/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2
urllib3==2.0.7
